Do not cache config struct
This fixes transactions being reported to the wrong apps.

The NewFangle::Config object was being cached on creation, but this
was breaking the possibility of generating multiple config objects
based on local configurations.

Hopefully, the app caching that is done in Plack::Middleware::NewFangle
will minimise the impact of this, and serve as a model of what type
of caching is needed for other implementations.