Use a local wrapper fork that can handle exceptions
This attempts to address https://github.com/cv-library/NewFangle-Agent/issues/2
by adding a fork of Hook::LexWrap that uses the new `defer` keyword
to ensure that the `post` wrappers are called even in the event of the
wrapped function dying.

Ideally, this would be added to the upstream code, but since this
would mean a pretty big behaviour change on a relatively old module,
chances are that merging that would be an uphill battle. While that
battle is fought, we can use this one and continue development on our
side.