Override global config with Defined environment variables
This makes it so that values in environment variables override the
config read during initialisation.

A similar change was added in 7717f4737bce6ee7550962fcbc0d5744387e0690
but that only applied to the config that was read via `local_config`.
With this change, the same behaviour applies to the config read via
`global_config`.